Moved all of my code into a lib.rs file
